bundle: reduce indirection in the parallel install scheduler
Addresses review feedback on comment volume and indirection.

`build_dependency_map`'s five single-use helpers are inlined back into it, per
the repo guideline to inline methods that are not reused twice or needed for
unit tests. None were called more than once and no spec targeted them
directly; the specs drive `build_dependency_map` itself, so coverage is
unchanged.

`Utils::StringTopologicalHash` is dropped and `Brew::Topo` restored exactly as
it was, so `utils/topological_hash.rb` and its spec are untouched by this
branch again. The two classes differed only in node type, `String` against
`CaskOrFormula`, which is not enough to justify a second class sitting in a
different file from either caller.

Comments in the changed hunks are cut to the reasons that are not evident from
the code. No behaviour change: every scheduling case this PR fixes produces
the same dependency map as before, and the attestation cases still match
`main` exactly.
