+++ Library/Homebrew/bundle/parallel_installer.rb
@@ -110,23 +110,87 @@ module Homebrew
 
       sig { params(entries: T::Array[Installer::InstallableEntry]).returns(T::Hash[String, T::Set[String]]) }
       def build_dependency_map(entries)
-        entry_deps = declared_deps_by_entry(entries)
-        lock_names = lock_names_by_entry(entries)
-        order = dependency_order(entries, entry_deps)
+        installed_taps = Homebrew::Bundle::Tap.installed_taps
+
+        # Map both full and short names so dep lookups work either way.
+        entry_name_map = entries.each_with_object(T.let({}, T::Hash[String, String])) do |entry, map|
+          map[entry.name] = entry.name
+          map[Utils.name_from_full_name(entry.name)] = entry.name
+        end
+
+        # Dependencies declared in the Brewfile, resolved to the entries providing them:
+        # entry A must finish before entry B starts.
+        entry_deps = entries.to_h do |entry|
+          deps = case entry.cls.name
+          when "Homebrew::Bundle::Brew"
+            Homebrew::Bundle::Brew.formula_dep_names(entry.name)
+          when "Homebrew::Bundle::Cask"
+            Homebrew::Bundle::Cask.formula_dependencies([entry.full_name]) +
+            Homebrew::Bundle::Cask.cask_dependencies([entry.full_name])
+          else
+            []
+          end
+          deps += Homebrew::Bundle::Installer.tap_dependencies(entry, entries:, installed_taps:)
+
+          resolved = deps.each_with_object(T.let(Set.new, T::Set[String])) do |dep, set|
+            name = entry_name_map[dep] || entry_name_map[Utils.name_from_full_name(dep)]
+            next if name.nil? || name == entry.name
+
+            set << name
+          end
+          [entry.name, resolved]
+        end
+
+        add_attestation_deps!(entries, entry_deps)
+
+        # `FormulaInstaller#lock` locks the formula and its recursive dependencies, so a
+        # formula with no dependencies of its own (e.g. `xz`) still conflicts with entries
+        # that depend on it. Cask installs take the same locks for their formulae.
+        lock_names = entries.to_h do |entry|
+          locks = case entry.cls.name
+          when "Homebrew::Bundle::Brew"
+            Homebrew::Bundle::Brew.lock_names(entry.name)
+          when "Homebrew::Bundle::Cask"
+            Homebrew::Bundle::Cask.lock_names([entry.full_name])
+          else
+            Set.new
+          end
+          [entry.name, locks]
+        end
+
+        # Conflicts must serialize the same way round as the declared dependencies, or a
+        # Brewfile listing a dependent first strands every entry in `run!`'s serial path.
+        topo = Homebrew::Bundle::Brew::Topo.new
+        entries.each { |entry| topo[entry.name] = entry_deps.fetch(entry.name).to_a }
+        order = topo.tsort_with_cycles do |cycles|
+          opoo "Installing these bundle entries one at a time, they depend on each " \
+               "other: #{cycles.flatten.uniq.join(", ")}"
+        end
         position = order.each_with_index.to_h
 
         # Formulae racing for an undeclared implicit dependency (e.g. a Linux sandbox
-        # executable) wait on just the first one, not on each other. It has to be the
-        # first in dependency order rather than in Brewfile order so that its edges
-        # point the same way round as everything else.
-        implicit_pioneer = if DependencyCollector.new.implicit_dependency_names.empty?
-          nil
-        else
+        # executable) wait on just the first one in dependency order, not on each other.
+        implicit_pioneer = unless DependencyCollector.new.implicit_dependency_names.empty?
           entries.select { |entry| entry.cls == Homebrew::Bundle::Brew }
-                 .min_by { |entry| position.fetch(entry.name) }&.name
+                 .min_by { |entry| position.fetch(entry.name, 0) }&.name
         end
 
-        merge_lock_conflicts(entries, entry_deps:, lock_names:, order:, position:, implicit_pioneer:)
+        entries.to_h do |entry|
+          depends_on = entry_deps.fetch(entry.name, Set.new).dup
+          entry_locks = lock_names.fetch(entry.name, Set.new)
+
+          order.take(position.fetch(entry.name, 0)).each do |earlier_name|
+            next if depends_on.include?(earlier_name)
+
+            depends_on << earlier_name if entry_locks.intersect?(lock_names.fetch(earlier_name, Set.new))
+          end
+
+          if implicit_pioneer && entry.name != implicit_pioneer && entry.cls == Homebrew::Bundle::Brew
+            depends_on << implicit_pioneer
+          end
+
+          [entry.name, depends_on]
+        end
       end
 
       sig { params(message: String, stream: IO).void }
@@ -144,11 +208,6 @@ module Homebrew
 
       private
 
-      sig { params(name: String).returns(String) }
-      def normalize_formula_name(name)
-        Utils.name_from_full_name(name)
-      end
-
       sig { params(entries: T::Array[Installer::InstallableEntry]).void }
       def prepare_attestation_verification!(entries)
         return unless Homebrew::EnvConfig.verify_attestations?
@@ -160,75 +219,8 @@ module Homebrew
         Homebrew::Attestation.gh_executable
       end
 
-      # The Cellar racks each entry's install locks. `FormulaInstaller#lock` locks the
-      # formula itself and all of its recursive dependencies before installing, even
-      # when pouring bottles, so a formula with no dependencies of its own (e.g. `xz`)
-      # still conflicts with every entry that depends on it. Cask installs take the
-      # same locks for the formulae they depend on.
-      sig { params(entries: T::Array[Installer::InstallableEntry]).returns(T::Hash[String, T::Set[String]]) }
-      def lock_names_by_entry(entries)
-        entries.to_h do |entry|
-          locks = case entry.cls.name
-          when "Homebrew::Bundle::Brew"
-            Homebrew::Bundle::Brew.lock_names(entry.name)
-          when "Homebrew::Bundle::Cask"
-            Homebrew::Bundle::Cask.lock_names([entry.full_name])
-          else
-            Set.new
-          end
-
-          [entry.name, locks]
-        end
-      end
-
-      # Dependencies declared in the Brewfile, resolved to the entries providing them.
-      # These determine install ordering: entry A must finish before entry B starts.
-      sig { params(entries: T::Array[Installer::InstallableEntry]).returns(T::Hash[String, T::Set[String]]) }
-      def declared_deps_by_entry(entries)
-        installed_taps = Homebrew::Bundle::Tap.installed_taps
-
-        # Map both full and short names so dep lookups work either way.
-        entry_name_map = entries.each_with_object(T.let({}, T::Hash[String, String])) do |entry, map|
-          map[entry.name] = entry.name
-          map[normalize_formula_name(entry.name)] = entry.name
-        end
-
-        entry_deps = entries.to_h do |entry|
-          deps = case entry.cls.name
-          when "Homebrew::Bundle::Brew"
-            Homebrew::Bundle::Brew.formula_dep_names(entry.name)
-          when "Homebrew::Bundle::Cask"
-            Homebrew::Bundle::Cask.formula_dependencies([entry.full_name]) +
-            Homebrew::Bundle::Cask.cask_dependencies([entry.full_name])
-          else
-            []
-          end
-
-          # Entries from non-default taps depend on the tap being installed first.
-          deps += Homebrew::Bundle::Installer.tap_dependencies(entry, entries:, installed_taps:)
-
-          resolved = deps.each_with_object(T.let(Set.new, T::Set[String])) do |dep, set|
-            name = entry_name_map[dep] || entry_name_map[normalize_formula_name(dep)]
-            next if name.nil? || name == entry.name
-
-            set << name
-          end
-
-          [entry.name, resolved]
-        end
-
-        add_attestation_deps!(entries, entry_deps)
-        entry_deps
-      end
-
-      # Attestation verification shells out to `gh`, so entries that verify one wait for
-      # it. `prepare_attestation_verification!` skips the upfront bootstrap when `gh` is
-      # itself a Brewfile entry, so this edge is the only thing sequencing it.
-      #
-      # Entries `gh` reaches through the declared dependencies above must be skipped,
-      # because that edge would point back against their own and close a cycle. Ask the
-      # graph the cycle would form in: a lock closure answers a different question, since
-      # it also covers build dependencies that carry no ordering edge.
+      # Entries verifying an attestation wait for a Brewfile `gh`, except those `gh`
+      # reaches itself: that edge would point back against their own and close a cycle.
       sig {
         params(
           entries:    T::Array[Installer::InstallableEntry],
@@ -256,58 +248,6 @@ module Homebrew
           next if entry.name == attestation_formula.name || reached.include?(entry.name)
 
           entry_deps.fetch(entry.name) << attestation_formula.name
-        end
-      end
-
-      # Orders the entries dependency-first. Lock conflicts have to be serialized the
-      # same way round as the declared dependencies, so a Brewfile that lists a
-      # dependent before its dependency (e.g. `python` before `xz`) does not end up
-      # with the two orderings disagreeing and stranding every entry in the sequential
-      # fallback in `run!`.
-      sig {
-        params(
-          entries:    T::Array[Installer::InstallableEntry],
-          entry_deps: T::Hash[String, T::Set[String]],
-        ).returns(T::Array[String])
-      }
-      def dependency_order(entries, entry_deps)
-        topo = ::Utils::StringTopologicalHash.new
-        entries.each { |entry| topo[entry.name] = entry_deps.fetch(entry.name).to_a }
-        # Entries outside a cycle still order correctly, so name the ones inside it and
-        # carry on. `Brew.sort!` already explains the usual cause (stale keg tab
-        # dependency data) and how to clear it, so do not repeat that advice here.
-        topo.sorted_names do |cycles|
-          opoo "Installing these bundle entries one at a time, they depend on each " \
-               "other: #{cycles.flatten.uniq.join(", ")}"
-        end
-      end
-
-      sig {
-        params(
-          entries:          T::Array[Installer::InstallableEntry],
-          entry_deps:       T::Hash[String, T::Set[String]],
-          lock_names:       T::Hash[String, T::Set[String]],
-          order:            T::Array[String],
-          position:         T::Hash[String, Integer],
-          implicit_pioneer: T.nilable(String),
-        ).returns(T::Hash[String, T::Set[String]])
-      }
-      def merge_lock_conflicts(entries, entry_deps:, lock_names:, order:, position:, implicit_pioneer:)
-        entries.to_h do |entry|
-          depends_on = entry_deps.fetch(entry.name, Set.new).dup
-          entry_locks = lock_names.fetch(entry.name, Set.new)
-
-          order.take(position.fetch(entry.name, 0)).each do |earlier_name|
-            next if depends_on.include?(earlier_name)
-
-            depends_on << earlier_name if entry_locks.intersect?(lock_names.fetch(earlier_name, Set.new))
-          end
-
-          if implicit_pioneer && entry.name != implicit_pioneer && entry.cls == Homebrew::Bundle::Brew
-            depends_on << implicit_pioneer
-          end
-
-          [entry.name, depends_on]
         end
       end
